src/services/puzzles/flow/audit/python/requirements.txt
# for using gRPC
grpcio
# for building and compiling gRPC clients/services
grpcio-tools
# a collection of generated python classes for some common protos
googleapis-common-protos